chore(): removes maincomponent overwrite and adds new dummy component instead
